+++ src/test/groovy/edu/ie3/datamodel/io/source/csv/CsvDataSourceTest.groovy
@@ -8,9 +8,11 @@ package edu.ie3.datamodel.io.source.csv
 import edu.ie3.datamodel.exceptions.SourceException
 import edu.ie3.datamodel.io.connectors.CsvFileConnector
 import edu.ie3.datamodel.io.csv.CsvIndividualTimeSeriesMetaInformation
+import edu.ie3.datamodel.io.csv.CsvLoadProfileMetaInformation
 import edu.ie3.datamodel.io.naming.FileNamingStrategy
 import edu.ie3.datamodel.io.naming.timeseries.ColumnScheme
 import edu.ie3.datamodel.models.input.system.LoadInput
+import edu.ie3.datamodel.models.profile.BdewStandardLoadProfile
 import spock.lang.Shared
 import spock.lang.Specification
 
@@ -51,6 +53,8 @@ class CsvDataSourceTest extends Specification implements CsvTestDataMeta {
   FileNamingStrategy fileNamingStrategy
   @Shared
   Set<Path> timeSeriesPaths
+  @Shared
+  Set<Path> loadProfileTimeSeriesPaths
 
   @Shared
   DummyCsvSource dummyCsvSource
@@ -70,6 +74,13 @@ class CsvDataSourceTest extends Specification implements CsvTestDataMeta {
       "its_c_c7b0d9d6-5044-4f51-80b4-f221d8b1f14b.csv"
     ].stream().map { file -> Path.of(file) }.collect(Collectors.toSet())
     timeSeriesPaths.forEach { path -> Files.createFile(testBaseFolderPath.resolve(path)) }
+
+    loadProfileTimeSeriesPaths = [
+      "lpts_r1.csv",
+      "lpts_r2.csv",
+      "lpts_g0.csv"
+    ].stream().map { file -> Path.of(file) }.collect(Collectors.toSet())
+    loadProfileTimeSeriesPaths.forEach { path -> Files.createFile(testBaseFolderPath.resolve(path)) }
   }
 
   def "A DataSource should contain a valid connector after initialization"() {
@@ -346,13 +357,17 @@ class CsvDataSourceTest extends Specification implements CsvTestDataMeta {
 
   def "The CsvDataSource is able to provide correct paths to time series files"() {
     when:
-    def actual = dummyCsvSource.getIndividualTimeSeriesFilePaths()
+    def actualIndividual = dummyCsvSource.getTimeSeriesFilePaths(fileNamingStrategy.individualTimeSeriesPattern)
+    def actualLoad = dummyCsvSource.getTimeSeriesFilePaths(fileNamingStrategy.loadProfileTimeSeriesPattern)
 
     then:
     noExceptionThrown()
 
-    actual.size() == timeSeriesPaths.size()
-    actual.containsAll(timeSeriesPaths)
+    actualIndividual.size() == timeSeriesPaths.size()
+    actualIndividual.containsAll(timeSeriesPaths)
+
+    actualLoad.size() == loadProfileTimeSeriesPaths.size()
+    actualLoad.containsAll(loadProfileTimeSeriesPaths)
   }
 
   def "The CsvDataSource is able to build correct uuid to meta information mapping"() {
@@ -388,5 +403,25 @@ class CsvDataSourceTest extends Specification implements CsvTestDataMeta {
 
     then:
     actual == expected
+  }
+
+  def "The CsvDataSource is able to build correct load profile meta information"() {
+    when:
+    def actual = dummyCsvSource.getCsvLoadProfileMetaInformation()
+
+    then:
+    actual.size() == 3
+    actual.get("r1").fullFilePath == Path.of("lpts_r1")
+    actual.get("r2").fullFilePath == Path.of("lpts_r2")
+    actual.get("g0").fullFilePath == Path.of("lpts_g0")
+  }
+
+  def "The CsvDataSource is able to build correct load profile meta information when restricting load profile"() {
+    when:
+    def actual = dummyCsvSource.getCsvLoadProfileMetaInformation(BdewStandardLoadProfile.G0)
+
+    then:
+    actual.size() == 1
+    actual.get("g0").fullFilePath == Path.of("lpts_g0")
   }
 }